ADDED: New Updates to APIs

+++ bs_cart_service/src/main/java/com/bridgelabz/bookstore/cart_service/services/CartService.java
@@ -6,8 +6,10 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import com.bridgelabz.bookstore.cart_service.dto.CartDto;
+import com.bridgelabz.bookstore.cart_service.model.BookModel;
 import com.bridgelabz.bookstore.cart_service.model.CartModel;
 import com.bridgelabz.bookstore.cart_service.repository.CartRepository;
 import com.bridgelabz.bookstore.cart_service.util.TokenUtil;
@@ -18,48 +20,75 @@ public class CartService implements ICartService {
 	TokenUtil tokenUtil;
 	@Autowired
 	CartRepository cartRepository;
+	@Autowired
+	RestTemplate restTemplate;
 
 	@Override
 	public CartModel addToCart(String token, CartDto cartDto) {
-		CartModel cart = new CartModel(cartDto);
-		cart.setUser_Id(tokenUtil.decodeToken(token));
-		return cartRepository.save(cart);
+		boolean verify = restTemplate.getForObject("http://BOOKSTORE-USER/bookstoreuser/verify?token=" + token,
+				Boolean.class);
+		if (verify == true) {
+			CartModel cart = new CartModel(cartDto);
+			BookModel book = restTemplate.getForObject("http://BOOK-CONTROLLER/bookstore/getbyid/"+cartDto.getBook_Id()+"?token="+token, BookModel.class);
+			System.out.println(book.toString());
+			if(cart.getBook_Id()== book.getBook_Id()) {
+			cart.setUser_Id(tokenUtil.decodeToken(token));
+			return cartRepository.save(cart);
+		}}
+		return null;
 
 	}
 
 	@Override
 	public List<CartModel> getAllCartItem(String token) {
-		List<CartModel> cartList = new ArrayList<CartModel>();
-		cartRepository.findAll().forEach(cartList::add);
-		return cartList;
+		boolean verify = restTemplate.getForObject("http://BOOKSTORE-USER/bookstoreuser/verify?token=" + token,
+				Boolean.class);
+		if (verify == true) {
+			List<CartModel> cartList = new ArrayList<CartModel>();
+			cartRepository.findAll().forEach(cartList::add);
+			return cartList;
+		}
+		return null;
 
 	}
 
 	@Override
-	public CartModel getCartItemsById(String token) {
-		Long id = tokenUtil.decodeToken(token);
-		return cartRepository.findById(id).get();
+	public CartModel getCartItemsById(String token, Long id) {
+		boolean verify = restTemplate.getForObject("http://BOOKSTORE-USER/bookstoreuser/verify?token=" + token,
+				Boolean.class);
+		if (verify == true) {
+			return cartRepository.findById(id).get();
+		}
+		return null;
 	}
 
 	@Override
 	public CartModel updateCartItem(String token, Long id, CartDto cartDto) {
-		Optional<CartModel> cartData = cartRepository.findById(id);
-		if (cartData.isPresent()) {
-			cartData.get().setBook_Id(cartDto.getBook_Id());
-			cartData.get().setQuantity(cartDto.getQuantity());
-			cartRepository.save(cartData.get());
-			return cartData.get();
+		boolean verify = restTemplate.getForObject("http://BOOKSTORE-USER/bookstoreuser/verify?token=" + token,
+				Boolean.class);
+		if (verify == true) {
+			Optional<CartModel> cartData = cartRepository.findById(id);
+			if (cartData.isPresent()) {
+				cartData.get().setBook_Id(cartDto.getBook_Id());
+				cartData.get().setQuantity(cartDto.getQuantity());
+				cartRepository.save(cartData.get());
+				return cartData.get();
+			}
 		}
 		return null;
 	}
 
 	@Override
 	public void removeCartItem(String token, Long id) {
-		Optional<CartModel> cartData = cartRepository.findById(id);
-		if (cartData.isPresent()) {
-			cartRepository.deleteById(id);
-		}
+		boolean verify = restTemplate.getForObject("http://BOOKSTORE-USER/bookstoreuser/verify?token=" + token,
+				Boolean.class);
+		if (verify == true) {
+			Optional<CartModel> cartData = cartRepository.findById(id);
+			if (cartData.isPresent()) {
+				cartRepository.deleteById(id);
+			}
 
+		}
 	}
 
 }
